iter2

+++ FileNode.java
@@ -1,14 +1,14 @@
 public class FileNode extends Node {
-    private int size; // 模拟字节数
+    private long size; // 模拟字节数
 
-    public FileNode(String name, int size) {
+    public FileNode(String name, long size) {
         super(name);
         if (size < 0) throw new IllegalArgumentException("size must be non-negative");
         this.size = size;
     }
 
     @Override
-    public int getSize() {
+    public long getSize() {
         return size;
     }
 }
